Update sambahsa.md
Sellamat Henrique ! 
I have corrected the grammatical points of the text in English up to line 7722 (then I had an issue that made disappear the text below. Please copy-paste or merge this "branch" with your version, so that Cale can start to work on a (relatively ?) grammatically correct text. (for what regards the grammar of Sambahsa)

Dank spollay !

Olivier